pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-assistant-competition"
version = "1.0.0"
description = "A learning platform for students to build AI agents that compete in negotiation challenges"
readme = "README.md"
requires-python = ">=3.9,<3.14"  # autogen requires Python < 3.14
license = {text = "MIT"}

dependencies = [
    "streamlit>=1.32.0",
    "psycopg2-binary",
    "pyjwt",
    "streamlit-aggrid",
    "autogen",
    "openai",
    "flask",
    "google-api-python-client",
    "google-auth",
    "pandas",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-playwright>=0.4.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0",
    "pre-commit>=3.0",
    "toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
markers = [
    "unit: Fast tests with no external dependencies (default)",
    "integration: Tests that may need mocked external services",
    "slow: Long-running tests",
    "requires_db: Tests that need a database connection",
    "requires_secrets: Tests that need secrets.toml",
    "e2e: End-to-end browser tests with Playwright",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["streamlit/modules"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
fail_under = 50
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py39"
line-length = 120
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "W291",   # trailing whitespace
    "W293",   # blank line contains whitespace
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]  # Allow assert in tests
"streamlit/pages/*" = ["E402"]  # Allow imports not at top (Streamlit pattern)
"streamlit/0_Home.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["streamlit.modules"]

[tool.black]
line-length = 120
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.venv
    | venv
    | __pycache__
    | \.git
    | build
    | dist
)/
'''

[tool.setuptools.packages.find]
where = ["."]
include = ["streamlit*"]
